requirements.txt
python-dotenv>=1.0.0
pyautogen>=0.2.0
typing-extensions>=4.7.1 